requirements.txt
absl-py==0.15.0
aiohttp @ file:///D:/bld/aiohttp_1637087223487/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1636093929600/work
antlr4-python3-runtime==4.8
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor @ file:///home/conda/feedstock_root/build_artifacts/astor_1593610464257/work
astunparse @ file:///home/conda/feedstock_root/build_artifacts/astunparse_1610696312422/work
async-timeout @ file:///home/conda/feedstock_root/build_artifacts/async-timeout_1640026696943/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1640799537051/work
Authlib==1.0.0
backcall==0.2.0
bcrypt==3.2.0
bleach==4.1.0
blinker==1.4
brotlipy @ file:///D:/bld/brotlipy_1636012392748/work
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1633010882559/work
certifi==2021.10.8
cffi @ file:///D:/bld/cffi_1636046278333/work
chain==1.0
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1643649795184/work
clang==5.0
click @ file:///D:/bld/click_1635822757640/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
cryptography @ file:///D:/bld/cryptography_1639699435873/work
cycler==0.11.0
Cython==0.29.25
debugpy @ file:///C:/ci/debugpy_1637091961445/work
decorator==4.4.2
deepcut==0.7.0.0
defusedxml==0.7.1
distlib==0.3.4
entrypoints==0.4
filelock==3.6.0
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
flatbuffers @ file:///home/conda/feedstock_root/build_artifacts/python-flatbuffers_1617723079010/work
fonttools==4.28.3
frozenlist @ file:///D:/bld/frozenlist_1643223056029/work
gast @ file:///home/conda/feedstock_root/build_artifacts/gast_1596839682936/work
google-api-core==2.7.1
google-api-python-client==2.41.0
google-auth @ file:///home/conda/feedstock_root/build_artifacts/google-auth_1629296548061/work
google-auth-httplib2==0.1.0
google-auth-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/google-auth-oauthlib_1630497468950/work
google-pasta==0.2.0
googleapis-common-protos==1.56.0
greenlet==1.1.2
grpcio @ file:///D:/bld/grpcio_1639715948696/work
h5py==3.1.0
httplib2==0.20.4
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1642433548627/work
importlib-metadata @ file:///D:/bld/importlib-metadata_1644556292918/work
iopath==0.1.9
ipykernel @ file:///C:/ci/ipykernel_1633545585502/work/dist/ipykernel-6.4.1-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1643818352265/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==7.6.5
itsdangerous==2.1.1
jedi @ file:///C:/ci/jedi_1644315428289/work
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-console==6.4.0
jupyter-core @ file:///C:/ci/jupyter_core_1636537172022/work
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.7.0
Keras-Preprocessing @ file:///home/conda/feedstock_root/build_artifacts/keras-preprocessing_1610713559828/work
kiwisolver==1.3.2
libclang==13.0.0
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1637220118004/work
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
multidict @ file:///D:/bld/multidict_1643055611394/work
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.19.5
oauthlib @ file:///home/conda/feedstock_root/build_artifacts/oauthlib_1643507977997/work
omegaconf==2.1.1
opencv-python==4.5.5.64
opt-einsum @ file:///home/conda/feedstock_root/build_artifacts/opt_einsum_1617859230218/work
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso @ file:///tmp/build/80754af9/parso_1596826841367/work
pathspec==0.9.0
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.4.0
platformdirs==2.5.1
plotly==5.6.0
portalocker==2.3.2
prometheus-client==0.13.1
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pydot==1.4.2
Pygments @ file:///tmp/build/80754af9/pygments_1600458456400/work
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1638819640841/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1643496850550/work
pyparsing==3.0.6
pyrsistent==0.18.1
PySocks @ file:///D:/bld/pysocks_1635862562854/work
pythainlp==3.0.3
python-crfsuite @ file:///D:/bld/python-crfsuite_1602439932710/work
python-dateutil==2.8.2
pytz==2021.3
pyu2f @ file:///home/conda/feedstock_root/build_artifacts/pyu2f_1604248910016/work
pywin32==302
pywinpty==2.0.2
PyYAML==6.0
pyzmq @ file:///C:/ci/pyzmq_1638435148211/work
qtconsole==5.2.2
QtPy==2.0.1
regex==2021.11.10
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1641580202195/work
requests-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/requests-oauthlib_1643557462909/work
rsa @ file:///home/conda/feedstock_root/build_artifacts/rsa_1637781155505/work
scikit-learn==1.0.1
scipy @ file:///C:/ci/scipy_1641555170412/work
seaborn==0.11.2
Send2Trash==1.8.0
six==1.15.0
sklearn==0.0
SQLAlchemy==1.4.32
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit @ file:///home/conda/feedstock_root/build_artifacts/tensorboard-plugin-wit_1641458951060/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorflow==2.6.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.0.0
tinydb==4.6.1
tk==0.1.0
toml==0.10.2
torch==1.10.0
tornado @ file:///C:/ci/tornado_1606924294691/work
tqdm==4.62.3
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
typing-extensions==3.7.4.3
uritemplate==4.1.1
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1641584929973/work
virtualenv==20.13.4
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1644332431572/work
widgetsnbextension==3.5.2
win-inet-pton @ file:///D:/bld/win_inet_pton_1635832858406/work
wincertstore==0.2
wrapt==1.12.1
WTForms==3.0.1
yacs==0.1.8
yarl @ file:///D:/bld/yarl_1636047033860/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1643828507773/work
